fix(insured-bridge): Changes to Arbitrum messenger contracts made after e2e testing

+++ packages/core/contracts/insured-bridge/ovm/Optimism_CrossDomainEnabled.sol
@@ -76,7 +76,7 @@ contract Optimism_CrossDomainEnabled {
      */
     function sendCrossDomainMessage(
         address _crossDomainTarget,
-        uint32 _gasLimit,
+        uint256 _gasLimit,
         bytes memory _message
     ) internal {
         getCrossDomainMessenger().sendMessage(_crossDomainTarget, _message, _gasLimit);
